backend-kms: Add support to track the active bo handles
The handle of EGL buffers are unnecessarily removed in use-cases where
multiple linux DMA buffers and EGL buffers are imported. The issue is
observed when same linux DMA buffers are imported multiple times.

To avoid the issue, add ref count tracing of the buffer object handles
used in libgbm.

Also it seems logical to move the allocated_handle flag to the structure
tracking buffer object handles.

Signed-off-by: Rajendraprasad K J <KarammelJayakumar.Rajendraprasad@in.bosch.com>